Migrating 'sqlite' tests to Mocha. Promptly commenting them out since the 'sqlite' package is no longer being maintained and does not build.
